Update README.md
Given the above list, create a program to iterate through all of the values and create a list with unique values and store the results to a list called Values2. The list called Values2 should contain every number that is in Values but NO DUPLICATES!  Display the results to the user. You must use a FOR LOOP!